+++ backend/src/core/services/upload.service.ts
@@ -13,6 +13,7 @@ import { jwtUtils } from '@core/utils/jwt'
 import { Campaign } from '@core/models'
 import { CsvStatusInterface } from '@core/interfaces'
 import { CSVParams } from '@core/types'
+import { StatsService, CampaignService } from '.'
 
 const MAX_PROCESSING_TIME = config.get('csvProcessingTimeout')
 
@@ -183,16 +184,52 @@ const checkTemplateKeysMatch = (
     throw new MissingTemplateKeysError(missingKeys)
   }
 }
+/**
+ * For campaign table, the s3 meta data is updated with the uploaded file, and its validity is set to true.
+ * update statistics with new unsent count
+ * @param param.transaction
+ * @param param.campaignId
+ * @param param.key
+ * @param param.filename
+ */
+const uploadCompleteOnComplete = ({
+  transaction,
+  campaignId,
+  key,
+  filename,
+}: {
+  transaction: Transaction
+  campaignId: number
+  key: string
+  filename: string
+}): ((numRecords: number) => Promise<void>) => {
+  return async (numRecords: number): Promise<void> => {
+    try {
+      // Updates metadata in project
+      await replaceCampaignS3Metadata(campaignId, key, filename, transaction)
+
+      await StatsService.setNumRecipients(campaignId, numRecords, transaction)
+
+      // Set campaign to valid
+      await CampaignService.setValid(campaignId, transaction)
+      transaction?.commit()
+    } catch (err) {
+      transaction?.rollback()
+      throw err
+    }
+  }
+}
 
 export const UploadService = {
   /*** S3 API Calls ****/
   getUploadParameters,
   extractParamsFromJwt,
   /**** Handle S3Key in DB *****/
-  replaceCampaignS3Metadata,
   storeS3TempFilename,
   storeS3Error,
   deleteS3TempKeys,
   getCsvStatus,
   checkTemplateKeysMatch,
+  /** Shared uploadComplete function */
+  uploadCompleteOnComplete,
 }
