+++ backend/src/email/services/email.service.ts
@@ -8,15 +8,13 @@ import {
   MailService,
   CampaignService,
   UploadService,
-  StatsService,
+  ProtectedService,
 } from '@core/services'
 import { MailToSend, CampaignDetails } from '@core/interfaces'
 
 import { EmailTemplate, EmailMessage } from '@email/models'
 import { EmailTemplateService } from '@email/services'
-import config from '@core/config'
-const PROTECTED_URL = config.get('protectedUrl')
-const PROTECT_METHOD_VERSION = 1
+
 /**
  * Gets a message's parameters
  * @param campaignId
@@ -220,46 +218,6 @@ const uploadCompleteOnChunk = ({
     }
   }
 }
-/**
- * For campaign table, the s3 meta data is updated with the uploaded file, and its validity is set to true.
- * update statistics with new unsent count
- * @param param.transaction
- * @param param.campaignId
- * @param param.key
- * @param param.filename
- */
-const uploadCompleteOnComplete = ({
-  transaction,
-  campaignId,
-  key,
-  filename,
-}: {
-  transaction: Transaction
-  campaignId: number
-  key: string
-  filename: string
-}): ((numRecords: number) => Promise<void>) => {
-  return async (numRecords: number): Promise<void> => {
-    try {
-      // Updates metadata in project
-      await UploadService.replaceCampaignS3Metadata(
-        campaignId,
-        key,
-        filename,
-        transaction
-      )
-
-      await StatsService.setNumRecipients(campaignId, numRecords, transaction)
-
-      // Set campaign to valid
-      await CampaignService.setValid(campaignId, transaction)
-      transaction?.commit()
-    } catch (err) {
-      transaction?.rollback()
-      throw err
-    }
-  }
-}
 
 const uploadProtectedCompleteOnPreview = ({
   transaction,
@@ -273,7 +231,6 @@ const uploadProtectedCompleteOnPreview = ({
   return async (data: CSVParams[]): Promise<void> => {
     // Checks the csv for all the necessary columns.
     const PROTECTED_CSV_HEADERS = ['recipient', 'payload', 'passwordhash', 'id']
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
     UploadService.checkTemplateKeysMatch(data, PROTECTED_CSV_HEADERS)
 
     EmailTemplateService.testHydration(
@@ -304,44 +261,12 @@ const uploadProtectedCompleteOnChunk = ({
 }): ((data: CSVParams[]) => Promise<void>) => {
   return async (data: CSVParams[]): Promise<void> => {
     try {
-      // These records go into the protected message table
-      const protectedMessages: Array<ProtectedMessageRecordInterface> = data.map(
-        (entry) => {
-          const { recipient, payload, passwordhash, id } = entry
-          return {
-            campaignId,
-            id,
-            recipient,
-            payload,
-            passwordHash: passwordhash,
-            version: PROTECT_METHOD_VERSION,
-          }
-        }
-      )
-      // START populate template
-      await ProtectedMessage.bulkCreate(protectedMessages, {
+      const messages = await ProtectedService.storeProtectedMessages({
         transaction,
-        logging: (_message, benchmark) => {
-          if (benchmark) {
-            logger.info(
-              `uploadProtectedCompleteOnChunk - ProtectedMessage: ElapsedTime ${benchmark} ms`
-            )
-          }
-        },
-        benchmark: true,
+        campaignId,
+        data,
       })
-
-      const emailMessages = protectedMessages.map(
-        ({ campaignId, recipient, id }) => ({
-          campaignId,
-          recipient,
-          params: {
-            recipient,
-            protectedlink: `${PROTECTED_URL}/${PROTECT_METHOD_VERSION}/${id}`,
-          },
-        })
-      )
-      await EmailMessage.bulkCreate(emailMessages, {
+      await EmailMessage.bulkCreate(messages, {
         transaction,
         logging: (_message, benchmark) => {
           if (benchmark) {
@@ -367,8 +292,6 @@ export const EmailService = {
   getHydratedMessage,
   uploadCompleteOnPreview,
   uploadCompleteOnChunk,
-  uploadCompleteOnComplete,
   uploadProtectedCompleteOnPreview,
   uploadProtectedCompleteOnChunk,
-  uploadProtectedCompleteOnComplete: uploadCompleteOnComplete,
 }
