+++ backend/src/telegram/services/telegram.service.ts
@@ -15,7 +15,7 @@ import { TelegramTemplateService } from '@telegram/services'
 
 import TelegramClient from './telegram-client.class'
 import { CSVParams } from '@core/types'
-import { UploadService, StatsService, CampaignService } from '@core/services'
+import { UploadService } from '@core/services'
 import logger from '@core/logger'
 
 /**
@@ -298,46 +298,6 @@ const uploadCompleteOnChunk = ({
     }
   }
 }
-/**
- * For campaign table, the s3 meta data is updated with the uploaded file, and its validity is set to true.
- * update statistics with new unsent count
- * @param param.transaction
- * @param param.campaignId
- * @param param.key
- * @param param.filename
- */
-const uploadCompleteOnComplete = ({
-  transaction,
-  campaignId,
-  key,
-  filename,
-}: {
-  transaction: Transaction
-  campaignId: number
-  key: string
-  filename: string
-}): ((numRecords: number) => Promise<void>) => {
-  return async (numRecords: number): Promise<void> => {
-    try {
-      // Updates metadata in project
-      await UploadService.replaceCampaignS3Metadata(
-        campaignId,
-        key,
-        filename,
-        transaction
-      )
-
-      await StatsService.setNumRecipients(campaignId, numRecords, transaction)
-
-      // Set campaign to valid
-      await CampaignService.setValid(campaignId, transaction)
-      transaction?.commit()
-    } catch (err) {
-      transaction?.rollback()
-      throw err
-    }
-  }
-}
 
 export const TelegramService = {
   findCampaign,
@@ -349,5 +309,4 @@ export const TelegramService = {
   validateAndConfigureBot,
   uploadCompleteOnPreview,
   uploadCompleteOnChunk,
-  uploadCompleteOnComplete,
 }
